Moved sms functionality to communication
